Refactor: Move UUIDs to headers and extract key logic to separate module
- Change UUID transmission from request body/query params to X-User-UUID header
- Create backend/keys.py module with all key management functionality:
  - File-based storage utilities (load_user_keys, save_user_keys, etc.)
  - API key validation functions (validate_anthropic_key, validate_github_key, validate_fly_key)
  - Helper functions (get_supported_providers, is_valid_provider)
- Refactor backend/app.py to use the new keys module
- Update frontend components to send UUIDs via headers:
  - SetupWindow.jsx: Send X-User-UUID header in POST requests
  - SetupContext.jsx: Send X-User-UUID header in GET requests
- Update tests to reflect new header-based UUID approach
- Add __pycache__/ to .gitignore
