+++ .gitignore
@@ -88,4 +88,4 @@ public
 
 # Temporary folders
 tmp/
-temp/
+temp/__pycache__/
